cv_service_trashnet/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
PyYAML==6.0.1
torch>=2.0.0
torchvision>=0.15.0
datasets==2.19.1
Pillow>=10.0.0
